Use object properties instead of arrays
This is a redesign of the library -- rather than storing the information
that will be converted to JSON in arrays on each object called
'properties', we store them in actual object properties.

Since the MessageCard format includes some properties that are
syntactically incorrect property names in PHP ("@type" and "@context"),
all the components of the card descend from `AbstractMessageCardEntity`,
which implements JsonSerializable and converts `$this->type` to
`"@type"`, for example.