Decouple MessageCard functionality from Monolog
Instead of having the TeamsHandler contain the curl logic for POSTing to
the O365 endpoint, let's put it in the MessageCard so that users don't
have to use Monolog.

Also it looks like MessageCards _must_ either have 'summary' or 'text'
properties, but the summary is not displayed by Teams (only by Outlook),
so by default we set the summary to the same thing as the card title.