Add excludedFields to TeamsHandler
We extend Monolog\Handler\AbstractHandler::pushProcessor so that
whenever the user pushes an AbstractPlaceholderProcessor, we keep track
of the placeholder in case it needs to be removed from the output at
some time in the future.

These placeholder values in `$record['extra']` will be removed from the
TeamsHandler default card, and can be accessed via
TeamsHandler::getExcludedFields.